Fix imports and refactor task mixins
